#1866 handled stats overview for officials
