+++ server/routes/statistics.js
@@ -51,8 +51,11 @@ var assign                     = require('es5-ext/object/assign')
   , getStatusHistory              = require('../mongo-queries/get-status-history')
   , getProcessingWorkingHoursTime = require('../../utils/get-processing-working-hours-time')
   , capitalize                    = require('es5-ext/string/#/capitalize')
-  , resolveFullStepPath  = require('../../utils/resolve-processing-step-full-path')
-  , _                    = require('mano').i18n
+  , resolveFullStepPath           = require('../../utils/resolve-processing-step-full-path')
+  , _                             = require('mano').i18n
+  , getTimeItemTemplate           = require('./utils/get-time-template')
+  , accumulateProcessingTimeItems = require('./utils/accumulate-processing-time-items')
+  , resolveTimePerPerson          = require('./utils/resolve-time-per-person')
   , processingStepsMetaWithoutFrontDesk =
 		require('../../utils/processing-steps-meta-without-front-desk')();
 
@@ -104,28 +107,6 @@ var businessProcessQueryHandler = new QueryHandler([{
 		});
 	}
 }]);
-
-var getTimeItemTemplate = function () {
-	return {
-		timedCount: 0,
-		maxTime: -Infinity,
-		minTime: Infinity,
-		totalTime: 0,
-		avgTime: 0
-	};
-};
-
-var accumulateProcessingTimeItems = function (collection, item) {
-	collection.timedCount++;
-	collection.totalTime += item.processingTime;
-	collection.minTime =
-		Math.min(collection.minTime, item.processingTime);
-	collection.maxTime =
-		Math.max(collection.maxTime, item.processingTime);
-	collection.avgTime =
-		Math.round(collection.totalTime
-			/ collection.timedCount);
-};
 
 module.exports = function (config) {
 	var driver = ensureDriver(ensureObject(config).driver)
@@ -247,94 +228,6 @@ module.exports = function (config) {
 
 			stepsResult.totalCorrectionsByUser = assign({}, stepsResult.totalCorrectionsByUser,
 				stepsResult.totalCorrections, { label: _("Corrections by the users") });
-
-			return stepsResult;
-		});
-	};
-
-	var resolveTimePerPerson = function (query) {
-		var stepsResult = {};
-		Object.keys(processingStepsMetaWithoutFrontDesk).forEach(function (stepShortPath) {
-			stepsResult[stepShortPath] = {};
-			stepsResult[stepShortPath].label = db['BusinessProcess' +
-				capitalize.call(processingStepsMeta[stepShortPath]._services[0])].prototype
-				.processingSteps.map.getBySKeyPath(resolveFullStepPath(stepShortPath)).label;
-			stepsResult[stepShortPath].rows = { totalProcessing: assign(getTimeItemTemplate(),
-				{ processor: _("Total & times") }) };
-		});
-
-		return getStatusHistory.find({
-			onlyFullItems: true,
-			dateFrom: query.dateFrom,
-			dateTo: query.dateTo,
-			service: query.service,
-			excludeFrontDesk: true,
-			sort: {
-				'service.businessName': 1,
-				'service.businessId': 1,
-				'date.ts': 1
-			}
-		}).then(function (statusHistory) {
-			var currentItem, step;
-			statusHistory.forEach(function (statusHistoryItem) {
-				if (statusHistoryItem.status.code === 'pending') {
-					currentItem = { bpId: statusHistoryItem.service.id };
-					currentItem.processingStart = statusHistoryItem.date.ts;
-				} else if (currentItem) {
-					if (!db[statusHistoryItem.service.type]) {
-						currentItem = null;
-						return;
-					}
-					step =
-						db[statusHistoryItem.service.type].prototype.resolveSKeyPath(
-							statusHistoryItem.processingStep.path
-						);
-					if (step && step.value) {
-						step = step.value;
-					} else {
-						currentItem = null;
-						return;
-					}
-					if (currentItem.processingStart > statusHistoryItem.date.ts
-							|| !currentItem.processingStart ||
-							currentItem.bpId !== statusHistoryItem.service.id) {
-						currentItem = null;
-						return;
-					}
-					currentItem.processingEnd = statusHistoryItem.date.ts;
-					currentItem.businessName = statusHistoryItem.service.businessName;
-					currentItem.processingTime =
-						getProcessingWorkingHoursTime(currentItem.processingStart, currentItem.processingEnd);
-					currentItem.processor = statusHistoryItem.operator.id;
-					var stepPath = step.key;
-					if (!stepsResult[stepPath]) { // child of group step
-						stepPath = step.owner.owner.owner.key + '/' + step.key;
-					}
-					if (!stepsResult[stepPath].rows[currentItem.processor]) {
-						stepsResult[stepPath].rows[currentItem.processor] = {
-							processing: assign(getTimeItemTemplate(), {
-								processor: statusHistoryItem.operator.name,
-								processingPeriods: []
-							})
-						};
-					}
-					stepsResult[stepPath].rows[currentItem.processor]
-						.processing.processingPeriods.push(currentItem);
-
-					accumulateProcessingTimeItems(
-						stepsResult[stepPath].rows[currentItem.processor].processing,
-						currentItem
-					);
-					accumulateProcessingTimeItems(stepsResult[stepPath].rows.totalProcessing, currentItem);
-
-					currentItem = null;
-				}
-			});
-			Object.keys(stepsResult).forEach(function (key) { // put totals in the end
-				var totalProcessing = stepsResult[key].rows.totalProcessing;
-				delete stepsResult[key].rows.totalProcessing;
-				stepsResult[key].rows.totalProcessing = totalProcessing;
-			});
 
 			return stepsResult;
 		});
