+++ server/routes/utils/get-time-template.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = function () {
+	return {
+		timedCount: 0,
+		maxTime: -Infinity,
+		minTime: Infinity,
+		totalTime: 0,
+		avgTime: 0
+	};
+};
